Refactor the `model/session` package
